+++ src/components/Menu.vue
@@ -100,6 +100,16 @@
                 ]"
             /></em>
           </li>
+          <li @click="showModernImages">
+          <small>Show Modern Images</small>
+          <em
+              ><font-awesome-icon
+                :icon="[
+                  'fas',
+                  grimoire.modernImages ? 'check-square' : 'square',
+                ]"
+            /></em>
+        </li>
           <li @click="toggleStatic">
             Disable Animations
             <em
@@ -296,6 +306,9 @@ export default {
         this.toggleImageOptIn();
       }
     },
+    showModernImages() {
+      this.toggleModernImages();
+    },
     joinSession() {
       if (this.session.sessionId) return this.leaveSession();
       let sessionId = prompt(
@@ -360,6 +373,7 @@ export default {
       "toggleGrimoire",
       "toggleMenu",
       "toggleImageOptIn",
+      "toggleModernImages",
       "toggleMuted",
       "toggleNightOrder",
       "toggleStatic",